GetDeviceContext() and FinishDeviceContext() and the usage of CmdBuffer in android::OpenXR to allow a CommandBuffer for each image in every swapchain. CmdBuffer::Wait() has been moved to GetDeviceContext() so as to not stall the CPU thread. Fix for PrintAt3dPos().

+++ pc_client/UseOpenXR.h
@@ -20,7 +20,7 @@ namespace teleport
 		}
 	protected:
 		const char *GetOpenXRGraphicsAPIExtensionName() const override;
-		platform::crossplatform::GraphicsDeviceContext& GetDeviceContext(int uint32_t) override
+		platform::crossplatform::GraphicsDeviceContext& GetDeviceContext(size_t, size_t) override
 		{
 			return *deviceContext;
 		}
